# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Andreuapp
  version: v0
  contact:
    email: a@a.ru
    name: John Doe
servers:
  - url: 'http://localhost:8080/'
    description: localhost
paths:
  /v1/groups:
    get:
      summary: Метод отображения всех групп
      tags:
        - Groups
      operationId: findAllGroups
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupReadDto'
        'default':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Создание или апдейт группы
      tags:
        - Groups
      operationId: createOrUpdateGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupEditDto'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupReadDto'
        'default':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'


  /v1/groups/{id}:
    get:
      summary: Метод получения группы по id
      tags:
        - Groups
      operationId: findGroupById
      parameters:
        - name: id
          in: path
          required: true
          description: Id группы
          schema:
            type: integer
            format: int32
          example: 2
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupReadDto'
        'default':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Удаление группы по id
      tags:
        - Groups
      operationId: deleteGroup
      parameters:
        - name: id
          in: path
          required: true
          description: Id группы
          schema:
            type: integer
            format: int32
          example: 2
      responses:
        '200':
          description: Успешный ответ
        'default':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/students:
    get:
      summary: Метод отображения количества студентов по группам
      tags:
        - Students
      operationId: findAllQuantity
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupCountDto'
        'default':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Апдейт группы у студента
      tags:
        - Students
      operationId: AddOrUpdatetoGroupStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpdateGroupDto'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentReadDto'
        'default':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/students/{id}:
    delete:
      summary: Удаление студента по id
      tags:
        - Students
      operationId: delete
      parameters:
        - name: id
          in: path
          required: true
          description: Id студента
          schema:
            type: integer
            format: int64
          example: 2
      responses:
        '200':
          description: Успешный ответ
        'default':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'


components:
  schemas:
    GroupReadDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        number:
          type: string
          example: "11-11"
        students:
          type: array
          items:
            $ref: '#/components/schemas/StudentReadDto'

    StudentReadDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        dateOfEnrollment:
          type: string
          format: date
          example: "1990-11-01"

    StudentUpdateGroupDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        number:
          type: string

    GroupCountDto:
      type: object
      properties:
        groupNum:
          type: string
        studentsCount:
          type: integer
          format: int64
      required:
        - groupNum
        - studentsCount
    GroupEditDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        number:
          type: string


    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
      required:
        - code
        - message